+++ aws_xray_sdk/ext/django/apps.py
@@ -35,6 +35,7 @@ class XRayConfig(AppConfig):
             service=settings.AWS_XRAY_TRACING_NAME,
             dynamic_naming=settings.DYNAMIC_NAMING,
             streaming_threshold=settings.STREAMING_THRESHOLD,
+            max_trace_back=settings.MAX_TRACE_BACK,
         )
 
         # if turned on subsegment will be generated on